brep(boolean): remove the redundant bespoke curved∩curved handlers (#1403)
The general SSI→trim→membership→stitch pipeline now builds every clean ruled
boolean the per-pair bespoke loop→body constructors used to: crossing-cylinder,
cone∩cone and cone∩cylinder (∩, −, ∪) and partial penetration. With those
migrations adopted (validBooleanSolid passes, OCC-validated, adoption-guarded),
the bespoke handlers were dead fallbacks — disabling all twelve left the full
kernel suite green — so they are removed and kernel/ops routes straight through
the general drivers (an out-of-scope pair declines to the planar CSG fallback in
booleanGeneral, exactly as the bespoke handlers did).

Deleted: the 11 bespoke assembly files (curved_crossing_{intersect,cut,join},
curved_cone_cone_{intersect,cut,join}, curved_cone_cylinder_{intersect,cut,join},
curved_partial_penetration, curved_cross_rod) and their 11 direct test files —
~2000 lines. Their behaviour is covered by the general-path tests, the OCC oracle,
and the adoption guards.

KEPT (still load-bearing, so NOT a clean file delete):
  - the imprint tracers (curved_*_imprint.go) — the general path's SSI stage;
  - pointInsideCylinderSolid + its frame helpers — the general path's cylinder
    membership oracle, moved to a new curved_cylinder_membership.go;
  - seamedCircle — moved to curved_steinmetz_cut.go, its sole remaining user;
  - the shared curved-boolean test helpers (assertWatertight/faceTypeCounts),
    moved to curved_boolean_test_helpers_test.go;
  - the equal-radius STEINMETZ handlers: the bicylinder's two ellipses pinch into
    four lobes, a self-intersecting imprint the (u,v) arrangement does not yet
    split, so Steinmetz stays on its dedicated analytic constructor.

Full kernel + model + app suites green (54 packages); lint/vet/gofmt clean.

Advances #1403.
