+++ kernel/ops/tolerance_guard_test.go
@@ -48,10 +48,12 @@ func TestNoUnjustifiedAbsoluteEpsilons(t *testing.T) {
 		"../brep/curved_halfspace_looped.go",
 		"../brep/curved_halfspace_cone_side.go",
 		"../brep/curved_crossing_imprint.go",
-		"../brep/curved_crossing_cut.go",
-		"../brep/curved_crossing_intersect.go",
-		"../brep/curved_cone_cylinder_intersect.go",
-		"../brep/curved_partial_penetration.go",
+		"../brep/curved_cone_cone_imprint.go",
+		"../brep/curved_cone_cylinder_imprint.go",
+		"../brep/curved_cylinder_membership.go",
+		"../brep/curved_general_boolean.go",
+		"../brep/curved_general_ruled_cutjoin.go",
+		"../brep/curved_general_wrapping_band.go",
 		"../brep/curved_subtract_prism.go",
 		"../brep/curved_coaxial_cylinder.go",
 		"../brep/curved_cylinder_boss.go",
